fix: update storybook setup
