+++ react-components/package.json
@@ -26,7 +26,7 @@
     "@babel/preset-react": "7.22.5",
     "@babel/preset-typescript": "7.22.5",
     "@cognite/cogs.js": "^9.17.0",
-    "@cognite/reveal": "^4.3.3",
+    "@cognite/reveal": "link:../viewer",
     "@cognite/sdk": "^8.2.0",
     "@storybook/addon-essentials": "7.0.22",
     "@storybook/addon-interactions": "7.0.22",
